Improved command parsing; requires translation of tab character
Parser now gets all character offsets properly. It turns out that
FRESS required the IMLAC to count all the characters of the
`!(0italic!)` codes, even though the characters are not visible.

Also line break counts as one character, not two.

The parser now expects the "tab" at the start of text lines, which
make matching easier. As tab is translated out, we translate it in
CMS. The ` character that we translate on input serves well for this, as
it's allowed by CP but never showed up on any Brown terminals of the
era.

CMS must be set to translate '\t' to '`'. The complete setup on
CMS is:

```
SET OUTPUT 05 `
SET INPUT ` 78SET INPUT ^ 5f
```

The input code mapping for `^` allows to enter the EBCDIC `¬`,
needed for keyword request strings.